v2.0
1. Force change URN and Invoice Number into integer type when transfer
data from raw data list to middle data list, so that in the final
output, these two column will not show extra "0"s in the template (for
example: 117 instead of 0000000117).
2. Force change Billing Amount into float.
3. Remove the process of producing middle csv file and AR file.
4. Added the function of copying CareSys Template into Resi Admin's
folder, feed the data into the template, and save it into any name that
the user entered, then delete the temp file.